requirements.txt
arabic-reshaper==2.1.3
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.9
cryptography==3.4.6
cssselect2==0.4.1
defusedxml==0.7.1
Django==2.1.15
django-betterforms==1.2
django-fernet-fields==0.6
django-pyodbc-azure==2.1.0.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==3.3
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
Pillow==8.1.0
pkg_resources==0.0.0
pycparser==2.20
PyMySQL==1.0.2
pyodbc==4.0.30
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2021.1
qrcode==6.1
reportlab==3.5.66
requests==2.26.0
six==1.15.0
stripe==2.63.0
tinycss2==1.1.0
urllib3==1.26.7
WeasyPrint==52.4
webencodings==0.5.1
xhtml2pdf==0.2.5
